Refactor invoker tests for reuse
